Don't flip the board if a player has no tiles
At the end of the game, we "flip" the board (not actually a flip) to
visually show which player has the most pieces.  This looks confusing if
one player lost by losing all his pieces, and is definitely not needed
to clarify who won, so let's skip this in that case.