Search to a constant depth
Towards the game the search depth jumps from 3/5/7 (easy/normal/hard) to
15/17/19, which takes several minutes to complete and causes the game to
hang during that time.  This is making it pretty much impossible to
finish a game of Iagno unless one player is completely eliminated before
that point.

Search instead to a constant depth of 3/5/7 throughout the game.  It
might make sense to go a bit deeper towards the end of the game when
playing on Hard, but 3/5/7 should work fairly well.

https://bugzilla.gnome.org/show_bug.cgi?id=668553