Improve computer player behavior
This change actually makes use of the around () and eval_heuristic ()
methods to improve the play of the AI player. At level 1, the behavior
is the same as before.

For testing I ran matches against the old computer player which was
playing LIGHT (both sides were at level 2). I ran a full game for each
of the 244 potentially unique openings of 8 pieces (before 8 pieces are
on the board, the computer just plays randomly).

When dark was played by the old code, its record was 100 - 141 - 3.
When dark was played by the new code, its record was 191 - 47- 6.

This is a significant improvement, and may turn out to be too difficult
of an opponent, especially at level 3. We could try reducing the search
depth to compensate... I'm not sure how to judge absolute difficulty of
the AI.

https://bugzilla.gnome.org/show_bug.cgi?id=708179