Add a delay between the last move and final flip
Currently, it often seems like the final player has placed a piece
of his opponent's color instead of his own, since after clicking the
piece that is shown placed on the board is the one that is fated to be
there after the final flip. This is confusing and made me think I was
seeing Bug #679137. Add a delay to prevent confusion.