Don't move extra fast on the last turn
I added this because I thought it would look weird for the AI to sit and
think about what its next move should be if there is only one unfilled
spot on the board, but this is so fast it's difficult to follow. There's
really no need to special-case anything; it's only two seconds.