Allow to see end position.
A minor change of behaviour,
even if we're in freeze. The
game ends with an animation,
for displaying the number of
tiles players have. A player
has two seconds to see final
position; that's not enough,
if the player wants to learn
by observing the game. Allow
reaching the end position by
the first call to an "undo".