Fix final animation bug.
When against an AI level 3
(probably also a level 2),
the end animation randomly
failed (a final-like board
was displayed with a false
count instead of the final
move, and then the correct
score was displayed). That
is hard to understand. But
it looks like this code is
fixing the bug, good news.