Introduce GameState.
A GameState object keeps all the
useful informations about a game
state. It is minimalistic and so
better when the AI deals with it
and will also be easily saved in
an history of the game not based
on a long list of flipped tiles.

Game is for now a subclass of it
but that should be changed next.