Highlight playable tiles.
That is done a quite obscure
way, as you have to click on
an occupied tile. But, it is
the only way I have found to
make keyboard and touch have
an acceptable behavior as it
is impossible to know if the
position is playable without
clicking on it (the hovering
highlight fixes this problem
when playing with mouse; but
I cannot find a great design
for touch, and using a "soft
highlight" also feels broken
when using keyboard, so that
will need ideas and design).